add rollback(); used when an exception occurs during import
During a rollback, issues created by the script will be either
deleted or closed, depending on the state of the given `--token`.

If `--token` is an administrator token, the script will delete issues
in order to rollback. If it's not an administrator, it will close issues
that were opened in the import process.

Signed-off-by: Kevin Morris <kevr@0cost.org>