gitlab: sudo as users if found on gitlab
Gitlab API has a 'sudo' feature to masquerade as users.
Now, we'll lower case both the input and gitlab usernames
we find, and attempt to match Gitlab users with the same
usernames of originating posters (tasks, comments).

If the user cannot be found on Gitlab, we'll fallback to
providing an annotation about them and optionally a link
to their user page when provided with an --upstream.

+ Fixed ordering bug for issue comments. GitLab ignores
  the 'created_at' field in certain circumstances, so
  we order them ourselves and supply a small annotation
  of the date that the comments were added.

Signed-off-by: Kevin Morris <kevr@0cost.org>