gitlab: support 'created_at' for group projects
This change adds support for 'created_at' to be properly set
in Gitlab's issues and notes. This requires that the user creating
the issue or note is of owner status in the group that contains
the project.

This commit does still allow for non-group project migrations,
resorting to the old behavior (including datetime as small text).

Signed-off-by: Kevin Morris <kevr@0cost.org>