...fix some clearly screwed up list indexing
This caused a fatal exception, of course, because we aren't
even indexing the same way.

The indexes in the task list are a result of the results from
a database fetch. The index we were looking for was a task id.

Just look at the task that we're focused on for the loop's
iteration instead of trying to re-index...

Signed-off-by: Kevin Morris <kevr@0cost.org>